+++ src/transpiler.ts
@@ -21,7 +21,6 @@ import {
   ExpressionSplitter,
   ExternalArgModifier,
   ExternalContractHandler,
-  FilePathMangler,
   FreeFunctionInliner,
   FunctionPruner,
   FunctionTypeStringMatcher,
@@ -105,7 +104,6 @@ function applyPasses(ast: AST, options: TranspilationOptions & PrintOptions): AS
     ['Tnr', TypeNameRemover],
     ['Ru', RejectUnsupportedFeatures],
     ['Wa', WarnSupportedFeatures],
-    ['Fm', FilePathMangler],
     ['Ss', SourceUnitSplitter],
     ['Ct', TypeStringsChecker],
     ['Ae', ABIExtractor],
